+++ tests/Feature/FilterTest.php
@@ -1,28 +1,19 @@
 <?php
 
-use Czernika\OrchidLogViewer\LogData;
-use Illuminate\Testing\TestResponse;
-
 uses()->group('feature.filter');
 
 describe('level filter', function () {
-    it('filters logs by level', function () {
-        $this->mockLogsWith([
-            new LogData([
-                'text' => 'Log ERROR message',
-                'level' => 'error',
-            ]),
-            new LogData([
-                'text' => 'Log INFO message',
-                'level' => 'info',
-            ]),
-        ]);
+    it('will not be shown at all if there are no logs')->todo();
 
-        /** @var TestResponse */
-        $response = $this->seeWithParameters([
-            'level' => 'info',
-        ]);
+    it('filters logs by level')->todo();
 
-        $response->assertDontSee('Log ERROR message');
-    })->todo();
+    it('filters logs by file')->todo();
+
+    it('will not show select level filter according to config option')->todo();
+
+    it('will not show select file filter according to config option')->todo();
+
+    it('shows default headings')->todo();
+
+    it('can change headings')->todo();
 });