+++ tests/Feature/LayoutTest.php
@@ -1,0 +1,7 @@
+<?php
+
+uses()->group('feature.layout');
+
+describe('custom layout', function () {
+    test('custom layout can be changed')->todo();
+});