+++ tests/Feature/LogServiceTest.php
@@ -1,0 +1,9 @@
+<?php
+
+uses()->group('feature.log-service');
+
+describe('log service', function () {
+    it('can delete selected file')->todo();
+
+    it('can clear selected file')->todo();
+});